Added new features / Bug fixed
・未入力の文字列項目を検索する機能を追加。
・RestoreLocalFilesパラメータ機能を追加。
・リンク一覧のソート時にビューで指定した表示項目が解除される問題を解消。
・IEで戻るボタン押下時の確認ダイアログでキャンセルするとJavaScriptのエラーが発生する問題を解消。
・数値以外の値でタイトル結合されたリンク項目の値を取得する問題を解消。
・計算式の同期が失敗する問題を解消。
・数値項目の一括更新が失敗する問題を解消。
・リンク項目のタイトル結合によりマスタレコード更新時のパフォーマンスが悪化する問題を解消。
・APIからの一括削除のフィルタが正しく機能しない問題を解消。
・.NET Framework版のコードとマージ。

+++ Implem.Pleasanter/Libraries/DataTypes/Attachment.cs
@@ -70,7 +70,7 @@ namespace Implem.Pleasanter.Libraries.DataTypes
                     "Attachments",
                     Guid));
             }
-            else if (Deleted == true)
+            else if (Deleted == true && !Parameters.BinaryStorage.RestoreLocalFiles)
             {
                 Files.DeleteFile(Path.Combine(
                     Directories.BinaryStorage(),
